Refactor location icons and update location management routes
 - Replace individual location icons with a new `LocationIcon` component.
 - Remove `account.locations.$locationId.edit.tsx` and integrate its functionality into `account.locations.$locationId._index.tsx`.
 - Add cache invalidation after location deletion in `account.locations.$locationId.destroy.tsx`.
 - Introduce new route `account.locations.$locationId.products.tsx` for managing products associated with a location.
 - Delete unused route `account.locations.$locationId.set-default.tsx`.
 - Adjust `account.locations._index.tsx` to link directly to the location ID route.
 - Include `originType` in the default values for service creation in `account.services.create.tsx`.
